rider template
