+++ frontend/src/styles/Map.css
@@ -1,9 +1,6 @@
 .map-container {
-  position: absolute;
-  left: 20%;
-  top: 0;
-  bottom: 0;
-  width: 500px;
+  position: relative;
+  width: 100%;
   height: 500px;
 
 }
